requirements.txt
opencv-python-headless
paho-mqtt
numpy
flask
